Bump version
In order to be able to repost the gem to RubyGems, after yanking it,
I need to bump the version.

This is kind of cumbersome, since the only change is really in the
description of the gem (there were \ breaking lines, but we should use
\n in specific spots of the text instead).

I hope this now gets formatted correctly, or else... version 0.0.3, lol.